[Chore] Factor out validators to separate module
Problem: wizard_structure contains too many definitions, which
became hard to maintain.

Solution: Put all validators under `validators` module. Strip
`_validator` from its functions, since they are supposed to be called
with module qualified. Add typing information to the `Validator` class.
